Update and minor refactor
