+++ src/lib.rs
@@ -122,407 +122,23 @@ pub use otelapis::opentelemetry;
 /// This module defines a skeleton implementation of the open telemetry
 /// collector tracing service
 ///
-pub mod trace {
-    use crate::opentelemetry::proto::collector::trace::v1 as base;
-    use crate::opentelemetry::proto::collector::trace::v1::trace_service_server as skel;
-
-    /// Alias tonic TraceRequest
-    pub type OtelTraceRequest = tonic::Request<base::ExportTraceServiceRequest>;
-
-    /// Alias tonic TraceResponse
-    pub type OtelTraceResponse = tonic::Response<base::ExportTraceServiceResponse>;
-
-    /// Alias the generated server skeletons
-    pub use skel::TraceServiceServer;
-
-    /// Alias trace callback fn
-    pub type OnTraceFn = dyn Fn(OtelTraceRequest) -> Result<OtelTraceResponse, tonic::Status>
-        + Send
-        + Sync
-        + 'static;
-
-    /// GRPC trace service skeleton
-    pub struct OtelTraceService {
-        on_trace: Box<OnTraceFn>,
-    }
-
-    impl OtelTraceService {
-        /// Creates a trace service with the specified trace event handler function
-        pub fn with_handler(handler: Box<OnTraceFn>) -> Self {
-            OtelTraceService { on_trace: handler }
-        }
-    }
-
-    /// Creates a tonic service handler for open telemetry trace events
-    pub fn make_service(handler: Box<OnTraceFn>) -> skel::TraceServiceServer<OtelTraceService> {
-        skel::TraceServiceServer::new(OtelTraceService::with_handler(handler))
-    }
-
-    #[tonic::async_trait]
-    impl skel::TraceService for OtelTraceService {
-        async fn export(
-            &self,
-            request: tonic::Request<base::ExportTraceServiceRequest>,
-        ) -> Result<tonic::Response<base::ExportTraceServiceResponse>, tonic::Status> {
-            (self.on_trace)(request)
-        }
-    }
-}
+pub mod trace;
 
 /// This module defines a skeleton implementation of the open telemetry
 /// collector logging service
 ///
 #[cfg(feature = "otel-logs")]
-pub mod logs {
-    use crate::opentelemetry::proto::collector::logs::v1 as base;
-    use crate::opentelemetry::proto::collector::logs::v1::logs_service_server as skel;
-    use async_channel::{Receiver, Sender};
-
-    /// Alias tonic request
-    pub type OtelLogsRequest = tonic::Request<base::ExportLogsServiceRequest>;
-
-    /// Alias tonic reponse
-    pub type OtelLogsResponse = tonic::Response<base::ExportLogsServiceResponse>;
-
-    /// Alias service skeleton
-    pub use skel::LogsService;
-
-    /// Alias logs server
-    pub use skel::LogsServiceServer;
-
-    /// Alias logs callback fn
-    pub type OnLogsFn =
-        dyn Fn(OtelLogsRequest) -> Result<OtelLogsResponse, tonic::Status> + Send + Sync + 'static;
-
-    /// GRPC logs service skeleton
-    pub struct OtelLogsService {
-        on_logs: Box<OnLogsFn>,
-    }
-
-    impl OtelLogsService {
-        /// Creates a logs service with the specified logs event handler function
-        pub fn with_handler(handler: Box<OnLogsFn>) -> Self {
-            OtelLogsService { on_logs: handler }
-        }
-    }
-
-    #[tonic::async_trait]
-    impl skel::LogsService for OtelLogsService {
-        async fn export(
-            &self,
-            request: tonic::Request<base::ExportLogsServiceRequest>,
-        ) -> Result<tonic::Response<base::ExportLogsServiceResponse>, tonic::Status> {
-            (self.on_logs)(request)
-        }
-    }
-
-    /// Creates a tonic service handler for open telemetry log events
-    pub fn make_service(handler: Box<OnLogsFn>) -> skel::LogsServiceServer<OtelLogsService> {
-        skel::LogsServiceServer::new(OtelLogsService::with_handler(handler))
-    }
-
-    /// Asynchronous channel sender
-    pub type OtelLogsSender = Sender<base::ExportLogsServiceRequest>;
-
-    /// Asynchronous channel receiver
-    pub type OtelLogsReceiver = Receiver<base::ExportLogsServiceRequest>;
-
-    /// Logs forwarding agent
-    pub struct OtelLogsServiceForwarder {
-        channel: Sender<base::ExportLogsServiceRequest>,
-    }
-
-    // Creates a metrics service with the specified asynchronous sender channel
-    impl OtelLogsServiceForwarder {
-        /// Creates a log forwarding agent with an asynchronous channel sender
-        pub fn with_sender(channel: Sender<base::ExportLogsServiceRequest>) -> Self {
-            OtelLogsServiceForwarder { channel }
-        }
-    }
-
-    #[tonic::async_trait]
-    impl skel::LogsService for OtelLogsServiceForwarder {
-        async fn export(
-            &self,
-            request: tonic::Request<base::ExportLogsServiceRequest>,
-        ) -> Result<tonic::Response<base::ExportLogsServiceResponse>, tonic::Status> {
-            match self.channel.send(request.into_inner()).await {
-                Ok(()) => Ok(tonic::Response::new(base::ExportLogsServiceResponse {
-                    partial_success: Some(base::ExportLogsPartialSuccess {
-                        rejected_log_records: 0,
-                        error_message: "Ok".to_string(),
-                    }),
-                })),
-                Err(e) => Err(tonic::Status::internal(format!(
-                    "Logs gRPC forwarder channel sender failed to dispatch {}",
-                    e
-                ))),
-            }
-        }
-    }
-
-    /// Creates a tonic service handler for open telemetry logs events
-    pub fn make_forwarder(
-        sender: OtelLogsSender,
-    ) -> skel::LogsServiceServer<OtelLogsServiceForwarder> {
-        skel::LogsServiceServer::new(OtelLogsServiceForwarder::with_sender(sender))
-    }
-}
+pub mod logs;
 
 /// This module defines a skeleton implementation of the open telemetry
 /// collector metrics service
 ///
 #[cfg(feature = "otel-metrics")]
-pub mod metrics {
-    use crate::opentelemetry::proto::collector::metrics::v1 as base;
-    use crate::opentelemetry::proto::collector::metrics::v1::metrics_service_server as skel;
-    use async_channel::{Receiver, Sender};
-
-    pub use skel::MetricsService;
-    pub use skel::MetricsServiceServer;
-
-    /// Alias tonic request
-    pub type OtelMetricsRequest = tonic::Request<base::ExportMetricsServiceRequest>;
-
-    /// Alias tonic response
-    pub type OtelMetricsResponse = tonic::Response<base::ExportMetricsServiceResponse>;
-
-    /// Alias metrics callback fn
-    pub type OnMetricsFn = dyn Fn(OtelMetricsRequest) -> Result<OtelMetricsResponse, tonic::Status>
-        + Send
-        + Sync
-        + 'static;
-
-    /// GRPC metrics service skeleton
-    pub struct OtelMetricsService {
-        on_metrics: Box<OnMetricsFn>,
-    }
-
-    impl OtelMetricsService {
-        /// Creates a metrics service with the specified metrics event handler function
-        pub fn with_handler(handler: Box<OnMetricsFn>) -> Self {
-            OtelMetricsService {
-                on_metrics: handler,
-            }
-        }
-    }
-
-    #[tonic::async_trait]
-    impl skel::MetricsService for OtelMetricsService {
-        async fn export(
-            &self,
-            request: tonic::Request<base::ExportMetricsServiceRequest>,
-        ) -> Result<tonic::Response<base::ExportMetricsServiceResponse>, tonic::Status> {
-            (self.on_metrics)(request)
-        }
-    }
-
-    /// Creates a tonic service handler for open telemetry metrics events
-    pub fn make_service(
-        handler: Box<OnMetricsFn>,
-    ) -> skel::MetricsServiceServer<OtelMetricsService> {
-        skel::MetricsServiceServer::new(OtelMetricsService::with_handler(handler))
-    }
-
-    /// Asynchronous channel sender
-    pub type OtelMetricsSender = Sender<base::ExportMetricsServiceRequest>;
-
-    /// Asynchronous channel receiver
-    pub type OtelMetricsReceiver = Receiver<base::ExportMetricsServiceRequest>;
-
-    /// Creates a metrics service with the specified asynchronous sender channel
-    pub struct OtelMetricsServiceForwarder {
-        channel: Sender<base::ExportMetricsServiceRequest>,
-    }
-
-    impl OtelMetricsServiceForwarder {
-        /// Creates a metrics service forwarding agent with an asynchronous channel sender
-        pub fn with_sender(channel: Sender<base::ExportMetricsServiceRequest>) -> Self {
-            OtelMetricsServiceForwarder { channel }
-        }
-    }
-
-    #[tonic::async_trait]
-    impl skel::MetricsService for OtelMetricsServiceForwarder {
-        async fn export(
-            &self,
-            request: tonic::Request<base::ExportMetricsServiceRequest>,
-        ) -> Result<tonic::Response<base::ExportMetricsServiceResponse>, tonic::Status> {
-            match self.channel.send(request.into_inner()).await {
-                Ok(_) => Ok(tonic::Response::new(base::ExportMetricsServiceResponse {
-                    partial_success: Some(base::ExportMetricsPartialSuccess {
-                        rejected_data_points: 0,
-                        error_message: "Ok".to_string(),
-                    }),
-                })),
-                Err(e) => Err(tonic::Status::internal(format!(
-                    "Metrics gRPC forwarder channel sender failed to dispatch {}",
-                    e
-                ))),
-            }
-        }
-    }
-
-    /// Creates a tonic service forwarder for open telemetry metrics events
-    pub fn make_forwarder(
-        sender: OtelMetricsSender,
-    ) -> skel::MetricsServiceServer<OtelMetricsServiceForwarder> {
-        skel::MetricsServiceServer::new(OtelMetricsServiceForwarder::with_sender(sender))
-    }
-}
+pub mod metrics;
 
 /// A unified set of services that provide log, metrics and trace events
 #[cfg(feature = "otel-all")]
-pub mod all {
-    use crate::opentelemetry::proto::collector::logs::v1 as logs_base;
-    use crate::opentelemetry::proto::collector::metrics::v1 as metrics_base;
-    use crate::opentelemetry::proto::collector::trace::v1 as trace_base;
-    use async_channel::{Receiver, Sender};
-    use std::net::SocketAddr;
-    //    use tonic::transport::Server;
-
-    /// Enumeration of protocol buffer messages that are sendable/receivable
-    pub enum OpenTelemetryEvents {
-        /// A logs export request
-        Logs(logs_base::ExportLogsServiceRequest, Option<SocketAddr>),
-        /// A metrics export request
-        Metrics(
-            metrics_base::ExportMetricsServiceRequest,
-            Option<SocketAddr>,
-        ),
-        /// A trace export request
-        Trace(trace_base::ExportTraceServiceRequest, Option<SocketAddr>),
-    }
-    impl From<tonic::Request<logs_base::ExportLogsServiceRequest>> for OpenTelemetryEvents {
-        fn from(req: tonic::Request<logs_base::ExportLogsServiceRequest>) -> Self {
-            let remote = req.remote_addr();
-            Self::Logs(req.into_inner(), remote)
-        }
-    }
-    impl From<tonic::Request<metrics_base::ExportMetricsServiceRequest>> for OpenTelemetryEvents {
-        fn from(req: tonic::Request<metrics_base::ExportMetricsServiceRequest>) -> Self {
-            let remote = req.remote_addr();
-            Self::Metrics(req.into_inner(), remote)
-        }
-    }
-    impl From<tonic::Request<trace_base::ExportTraceServiceRequest>> for OpenTelemetryEvents {
-        fn from(req: tonic::Request<trace_base::ExportTraceServiceRequest>) -> Self {
-            let remote = req.remote_addr();
-            Self::Trace(req.into_inner(), remote)
-        }
-    }
-
-    /// Alias receiver
-    pub type OpenTelemetrySender = Sender<OpenTelemetryEvents>;
-
-    /// Alias sender
-    pub type OpenTelemetryReceiver = Receiver<OpenTelemetryEvents>;
-
-    /// Creates a logs service with the specified asynchronous sender channel
-    pub struct LogsServiceForwarder {
-        channel: Sender<OpenTelemetryEvents>,
-    }
-
-    impl LogsServiceForwarder {
-        /// Creates a logs service forwarding agent
-        pub fn with_sender(channel: Sender<OpenTelemetryEvents>) -> Self {
-            LogsServiceForwarder { channel }
-        }
-    }
-
-    #[tonic::async_trait]
-    impl super::logs::LogsService for LogsServiceForwarder {
-        async fn export(
-            &self,
-            request: tonic::Request<logs_base::ExportLogsServiceRequest>,
-        ) -> Result<tonic::Response<logs_base::ExportLogsServiceResponse>, tonic::Status> {
-            match self.channel.send(OpenTelemetryEvents::from(request)).await {
-                Ok(_) => Ok(tonic::Response::new(logs_base::ExportLogsServiceResponse {
-                    partial_success: Some(logs_base::ExportLogsPartialSuccess {
-                        rejected_log_records: 0,
-                        error_message: "Ok".to_string(),
-                    }),
-                })),
-                Err(e) => Err(tonic::Status::internal(format!(
-                    "Logs gRPC forwarder channel sender failed to dispatch {}",
-                    e
-                ))),
-            }
-        }
-    }
-
-    /// Creates a metrics service with the specified asynchronous sender channel
-    pub struct MetricsServiceForwarder {
-        channel: Sender<OpenTelemetryEvents>,
-    }
-
-    impl MetricsServiceForwarder {
-        /// Creates a metrics service forwarding agent
-        pub fn with_sender(channel: Sender<OpenTelemetryEvents>) -> Self {
-            MetricsServiceForwarder { channel }
-        }
-    }
-
-    #[tonic::async_trait]
-    impl super::metrics::MetricsService for MetricsServiceForwarder {
-        async fn export(
-            &self,
-            request: tonic::Request<metrics_base::ExportMetricsServiceRequest>,
-        ) -> Result<tonic::Response<metrics_base::ExportMetricsServiceResponse>, tonic::Status>
-        {
-            match self.channel.send(OpenTelemetryEvents::from(request)).await {
-                Ok(_) => Ok(tonic::Response::new(
-                    metrics_base::ExportMetricsServiceResponse {
-                        partial_success: Some(metrics_base::ExportMetricsPartialSuccess {
-                            rejected_data_points: 0,
-                            error_message: "snot".to_string(),
-                        }),
-                    },
-                )),
-                Err(e) => Err(tonic::Status::internal(format!(
-                    "Metrics gRPC forwarder channel sender failed to dispatch {}",
-                    e
-                ))),
-            }
-        }
-    }
-
-    /// Creates a trace service with the specified asynchronous sender channel
-    pub struct TraceServiceForwarder {
-        channel: Sender<OpenTelemetryEvents>,
-    }
-
-    impl TraceServiceForwarder {
-        /// Creates a trace service forwarding agent
-        pub fn with_sender(channel: Sender<OpenTelemetryEvents>) -> Self {
-            TraceServiceForwarder { channel }
-        }
-    }
-
-    #[tonic::async_trait]
-    impl trace_base::trace_service_server::TraceService for TraceServiceForwarder {
-        async fn export(
-            &self,
-            request: tonic::Request<trace_base::ExportTraceServiceRequest>,
-        ) -> Result<tonic::Response<trace_base::ExportTraceServiceResponse>, tonic::Status>
-        {
-            match self.channel.send(OpenTelemetryEvents::from(request)).await {
-                Ok(_) => Ok(tonic::Response::new(
-                    trace_base::ExportTraceServiceResponse {
-                        partial_success: Some(trace_base::ExportTracePartialSuccess {
-                            rejected_spans: 0,
-                            error_message: "snot".to_string(),
-                        }),
-                    },
-                )),
-                Err(e) => Err(tonic::Status::internal(format!(
-                    "Trace gRPC forwarder channel sender failed to dispatch {}",
-                    e
-                ))),
-            }
-        }
-    }
-}
+pub mod all;
 
 #[cfg(test)]
 mod test {
